cli: move oidc token into pytest
