+++ Makefile
@@ -7,10 +7,10 @@ ALL_PY_SRCS := action.py \
 all:
 	@echo "Run my targets individually!"
 
-env/pyvenv.cfg: dev-requirements.txt
+env/pyvenv.cfg: dev-requirements.txt requirements.txt
 	python3 -m venv env
 	./env/bin/python -m pip install --upgrade pip
-	./env/bin/python -m pip install --requirement dev-requirements.txt
+	./env/bin/python -m pip install --requirement dev-requirements.txt --requirement requirements.txt
 
 .PHONY: dev
 dev: env/pyvenv.cfg
