+++ test/conftest.py
@@ -1,10 +1,15 @@
 import os
 import shutil
 import tempfile
+import time
+from datetime import datetime, timedelta
+from io import BytesIO
 from pathlib import Path
-from typing import Callable, Tuple, TypeVar
+from typing import Callable, Optional, Tuple, TypeVar
+from zipfile import ZipFile
 
 import pytest  # type: ignore
+import requests
 
 from .client import (BundleMaterials, SignatureCertificateMaterials,
                      SigstoreClient, VerificationMaterials)
@@ -12,6 +17,16 @@ from .client import (BundleMaterials, SignatureCertificateMaterials,
 _M = TypeVar("_M", bound=VerificationMaterials)
 _MakeMaterialsByType = Callable[[str, _M], Tuple[Path, _M]]
 _MakeMaterials = Callable[[str], Tuple[Path, VerificationMaterials]]
+
+_OIDC_BEACON_API_URL = (
+    "https://api.github.com/repos/sigstore-conformance/extremely-dangerous-public-oidc-beacon/"
+    "actions"
+)
+_OIDC_BEACON_WORKFLOW_ID = 55399612
+
+
+class OidcTokenError(Exception):
+    pass
 
 
 def pytest_addoption(parser):
@@ -54,12 +69,96 @@ def pytest_configure(config):
 
 
 @pytest.fixture
-def client(pytestconfig):
+def identity_token(pytestconfig):
+    gh_token = os.getenv("GHA_SIGSTORE_GITHUB_TOKEN")
+    if gh_token is None:
+        raise OidcTokenError(
+            "`GHA_SIGSTORE_GITHUB_TOKEN` environment variable not found"
+        )
+
+    session = requests.Session()
+    headers = {
+        "Accept": "application/vnd.github+json",
+        "X-GitHub-Api-Version": "2022-11-28",
+        "Authorization": f"Bearer {gh_token}",
+    }
+
+    workflow_time: Optional[datetime] = None
+    run_id: str
+
+    # We need a token that was generated in the last 5 minutes. Keep checking until we find one.
+    while workflow_time is None or datetime.now() - workflow_time >= timedelta(
+        minutes=5
+    ):
+        # If there's a lot of traffic in the GitHub Actions cron queue, we might not have a valid
+        # token to use. In that case, wait for 30 seconds and try again.
+        if workflow_time is not None:
+            # FIXME(jl): logging in pytest?
+            # _log("Couldn't find a recent token, waiting...")
+            time.sleep(30)
+
+        resp: requests.Response = session.get(
+            url=_OIDC_BEACON_API_URL + f"/workflows/{_OIDC_BEACON_WORKFLOW_ID}/runs",
+            headers=headers,
+        )
+        resp.raise_for_status()
+
+        resp_json = resp.json()
+        workflow_runs = resp_json["workflow_runs"]
+        if not workflow_runs:
+            raise OidcTokenError(f"Found no workflow runs: {resp_json}")
+
+        workflow_run = workflow_runs[0]
+
+        # If the job is still running, the token artifact won't have been generated yet.
+        if workflow_run["status"] != "completed":
+            continue
+
+        run_id = workflow_run["id"]
+        workflow_time = datetime.strptime(
+            workflow_run["run_started_at"], "%Y-%m-%dT%H:%M:%SZ"
+        )
+
+    resp = session.get(
+        url=_OIDC_BEACON_API_URL + f"/runs/{run_id}/artifacts",
+        headers=headers,
+    )
+    resp.raise_for_status()
+
+    resp_json = resp.json()
+    artifacts = resp_json["artifacts"]
+    if len(artifacts) != 1:
+        raise OidcTokenError(
+            f"Found unexpected number of artifacts on OIDC beacon run: {artifacts}"
+        )
+
+    oidc_artifact = artifacts[0]
+    if oidc_artifact["name"] != "oidc-token":
+        raise OidcTokenError(
+            f"Found unexpected artifact on OIDC beacon run: {oidc_artifact['name']}"
+        )
+    artifact_id = oidc_artifact["id"]
+
+    # Download the OIDC token artifact and unzip the archive.
+    resp = session.get(
+        url=_OIDC_BEACON_API_URL + f"/artifacts/{artifact_id}/zip",
+        headers=headers,
+    )
+    resp.raise_for_status()
+
+    with ZipFile(BytesIO(resp.content)) as artifact_zip:
+        artifact_file = artifact_zip.open("oidc-token.txt")
+
+        # Strip newline.
+        return artifact_file.read().decode().rstrip()
+
+
+@pytest.fixture
+def client(pytestconfig, identity_token):
     """
     Parametrize each test with the client under test.
     """
     entrypoint = pytestconfig.getoption("--entrypoint")
-    identity_token = pytestconfig.getoption("--identity-token")
     return SigstoreClient(entrypoint, identity_token)
